% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaphonebr.R
\name{capitalize_remove_accents}
\alias{capitalize_remove_accents}
\title{Phonetic preprocessing: removes accents, numbers and capitalizes}
\usage{
capitalize_remove_accents(fullname)
}
\arguments{
\item{fullname}{a character vector.}
}
\value{
a preprocessed character vector.
}
\description{
Remove diacritics, capitalizes and remove
characters that are not letters or spaecs.
}
\examples{
metaphonebr:::capitalize_remove_accents("Jo\u00e3o Silva 123!")
}
\keyword{internal}
