% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaphonebr-package.R
\docType{package}
\name{metaphonebr-package}
\alias{metaphonebr-package}
\title{metaphonebr: Custom 'MetaphoneBR' Phonetic Encoding for Brazilian Names}
\description{
Simplifies Brazilian names phonetically using a custom 'metaphoneBR' algorithm that preserves ending vowels. Useful for name matching processing preserving gender information carried generally by ending vowels in Portuguese. Mation (2025) \doi{10.6082/uchicago.15104}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ipeadata-lab/metaphonebr}
  \item \url{https://ipeadata-lab.github.io/metaphonebr/}
  \item Report bugs at \url{https://github.com/ipeadata-lab/metaphonebr/issues}
}

}
\author{
\strong{Maintainer}: Rodrigo Borges \email{rodrigoesborges@gmail.com} (\href{https://orcid.org/0000-0003-2076-1424}{ORCID})

Authors:
\itemize{
  \item Pedro Cavalcanti G. Ferreira (\href{https://orcid.org/0000-0002-8540-1860}{ORCID})
  \item Lucas Mation (\href{https://orcid.org/0000-0002-7461-932X}{ORCID})
}

Other contributors:
\itemize{
  \item Ipea - Institue for Applied Economic Research [copyright holder, funder]
}

}
\keyword{internal}
