% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaphonebr.R
\name{remove_dup_letters_spaces}
\alias{remove_dup_letters_spaces}
\title{Phonetic Removal: duplicated letters and spaces}
\usage{
remove_dup_letters_spaces(fullname)
}
\arguments{
\item{fullname}{A character vector.}
}
\value{
A character vector with no repeated letters nor spaces.
}
\description{
Remove duplicated letters and spaces.
}
\examples{
metaphonebr:::remove_dup_letters_spaces(c("  CARRO OSSO  ", "JOAOZINHO"))
}
\keyword{internal}
