% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaphonebr.R
\name{remove_duplicated_vowels}
\alias{remove_duplicated_vowels}
\title{Phonetic Removal: Repeated Vowels}
\usage{
remove_duplicated_vowels(fullname)
}
\arguments{
\item{fullname}{A character vector.}
}
\value{
A character vector with duplicated vowels removed.
}
\description{
Compress adjacent identical vowel sequences.
}
\examples{
# metaphonebr:::remove_duplicated_vowels(c("AARAO", "REEBECA"))
}
\keyword{internal}
