% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaphonebr.R
\name{remove_silent_letters}
\alias{remove_silent_letters}
\title{Phonetic Simplification: removal of silent letters}
\usage{
remove_silent_letters(fullname)
}
\arguments{
\item{fullname}{a character vector.}
}
\value{
a character vector with silent initial 'H's removed.
}
\description{
Removes silent 'H'  at the beggining of each word.
}
\examples{
metaphonebr:::remove_silent_letters("Helena Horta")
}
\keyword{internal}
