% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaphonebr.R
\name{simplify_consonants}
\alias{simplify_consonants}
\title{Phonetic Simplification: Similar Consonants}
\usage{
simplify_consonants(fullname)
}
\arguments{
\item{fullname}{A character vector.}
}
\value{
A character vector with simplified consonants.
}
\description{
Represent similar consonants with single representation.
}
\examples{
# metaphonebr:::simplify_consonants("CA\\u00c7ADOR CELSO CARLOS GEOVANIA WALTER YARA ZEBRA")
}
\keyword{internal}
