% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaphonebr.R
\name{simplify_digraphs}
\alias{simplify_digraphs}
\title{Phonetic Simplification: similar digraphs}
\usage{
simplify_digraphs(fullname)
}
\arguments{
\item{fullname}{a character vector.}
}
\value{
a character vector with simplified representation of digraphs.
}
\description{
Transforms common sounding digraphs to simplify their phonetic representation.
}
\examples{
metaphonebr:::simplify_digraphs("FILHA MANHA CHICO SCHMIDT SCENA ESCOVA QUILO")
}
\keyword{internal}
