% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaphonebr.R
\name{simplify_ending_nasals}
\alias{simplify_ending_nasals}
\title{Phonetic Simplification: Ending Nasal Sounds}
\usage{
simplify_ending_nasals(fullname)
}
\arguments{
\item{fullname}{A character vector.}
}
\value{
A character vector with simplified nasal sounds.
}
\description{
Unifies Ending Nasal Sounds.
}
\examples{
metaphonebr:::simplify_ending_nasals(c("JOAQUIN", "JOAQUIM"))
}
\keyword{internal}
