% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaphonebr-package.R
\docType{package}
\name{metaphonebr-package}
\alias{metaphonebr-package}
\title{metaphonebr: Simplifies Brazilian Names Phonetically with Custom MetaphoneBR algorithm}
\description{
Simplifies brazilian names phonetically using a custom metaphoneBR algorithm that preserves ending vowels.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ipeadata-lab/metaphonebr}
  \item \url{https://ipeadata-lab.github.io/metaphonebr/}
}

}
\author{
\strong{Maintainer}: Rodrigo Borges \email{rodrigoesborges@gmail.com} (\href{https://orcid.org/0000-0003-2076-1424}{ORCID})

Authors:
\itemize{
  \item Pedro Cavalcanti G. Ferreira (\href{https://orcid.org/0000-0002-8540-1860}{ORCID})
  \item Lucas Mation (\href{https://orcid.org/0000-0002-7461-932X}{ORCID})
}

}
\keyword{internal}
